backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.86.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.2",
    "mcp>=1.0.0",
    "passlib[bcrypt]>=1.7",
    "pydantic-settings>=2.0",
    "pyjwt[crypto]>=2.8",
    "python-multipart>=0.0.9",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.42.0",
]

[project.optional-dependencies]
voice = ["faster-whisper>=1.1.1"]

[tool.pyright]
enableTypeIgnoreComments = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportUnnecessaryComparison = true
reportUnnecessaryContains = true
reportUnnecessaryIsInstance = true
reportUnnecessaryTypeIgnoreComment = true
typeCheckingMode = "standard"

[dependency-groups]
dev = ["pyright>=1.1.408", "ruff>=0.15.8"]
